start adding TcmService in Azure. Organized solution

+++ SpeedOptimizer.Tests/PowerMapTests.cs
@@ -18,8 +18,8 @@ namespace SpeedOptimizer.Tests
                 { 700, 800, 900, 0, 5 },
                 { 700, 800, 900, 2, -2 }
             };
-            double[] speeds = { 0, 0.5,1, 2, 4};
-            double[] tractionPercentages = { 0, 0.5, 1, 2, 6};
+            float[] speeds = { 0, 0.5f,1, 2, 4};
+            float[] tractionPercentages = { 0, 0.5f, 1, 2, 6};
 
             // Act
             PowerMap powerMap = new PowerMap(powers, speeds, tractionPercentages);
@@ -39,8 +39,8 @@ namespace SpeedOptimizer.Tests
                 { 400, 500, 600 },
                 { 700, 800, 900 }
             };
-            double[] speeds = { 0, 2, 3 };
-            double[] tractionPercentages = { 0, 2, 3 };
+            float[] speeds = { 0, 2, 3 };
+            float[] tractionPercentages = { 0, 2, 3 };
 
             // Act
             PowerMap powerMap = new PowerMap(powers, speeds, tractionPercentages);
